refactored gallery item components
