+++ src/components/GalleryItem/GalleryItem.types.ts
@@ -1,6 +1,6 @@
 import GalleryItemMaterial from "../../core/gallery/GalleryItemMaterial";
 import { PropsWithChildren } from "react";
-import { Color, Material, Mesh, Texture } from "three";
+import { Material, Mesh } from "three";
 
 export const GALLERY_ITEM_NO_PROVIDER_FLAG = Symbol("GALLERY_ITEM_NO_PROVIDER");
 
@@ -18,20 +18,3 @@ export type GalleryItemProps = PropsWithChildren<{
   itemMaterial: GalleryItemMaterial;
   onInit?: (data: GalleryItemInitData) => void;
 }>;
-
-export type SolidColorItemProps = Omit<GalleryItemProps, "itemMaterial"> & {
-  color: string | Color;
-};
-
-export type ImageItemProps = Omit<GalleryItemProps, "itemMaterial"> & {
-  src?: string;
-  texture?: Texture;
-};
-
-export type VideoItemProps = Omit<GalleryItemProps, "itemMaterial"> & {
-  src: string;
-  autoplay?: boolean;
-  muted?: boolean;
-  loop?: boolean;
-  crossOrigin?: "anonymous" | "use-credentials" | "" | null;
-};
