+++ src/components/GalleryItem/index.ts
@@ -1,6 +1,3 @@
-export { default as GalleryItem } from "./GalleryItem";
-export { default as SolidColorItem } from "./SolidColorItem";
-export { default as ImageItem } from "./ImageItem";
-export { default as VideoItem } from "./VideoItem";
-export { default as usePlacementOnGalleryItem } from "./usePlacementOnGalleryItem";
+export * from "./GalleryItem";
+export * from "./usePlacementOnGalleryItem";
 export * from "./GalleryItem.types";
