+++ src/components/ImageItem/ImageItem.types.ts
@@ -1,0 +1,7 @@
+import { Texture } from "three";
+import { GalleryItemProps } from "../GalleryItem";
+
+export type ImageItemProps = Omit<GalleryItemProps, "itemMaterial"> & {
+  src?: string;
+  texture?: Texture;
+};
