+++ src/components/ImageItem/index.ts
@@ -1,0 +1,2 @@
+export * from "./ImageItem";
+export * from "./ImageItem.types";
