+++ src/components/SolidColorItem/SolidColorItem.types.ts
@@ -1,0 +1,6 @@
+import { Color } from "three";
+import { GalleryItemProps } from "../GalleryItem";
+
+export type SolidColorItemProps = Omit<GalleryItemProps, "itemMaterial"> & {
+  color: string | Color;
+};
