+++ src/components/SolidColorItem/index.ts
@@ -1,0 +1,2 @@
+export * from "./SolidColorItem";
+export * from "./SolidColorItem.types";
