+++ src/components/VideoItem/index.ts
@@ -1,0 +1,2 @@
+export * from "./VideoItem";
+export * from "./VideoItem.types";
