+++ src/components/index.ts
@@ -2,3 +2,6 @@ export * from "./Ground";
 export * from "./Gallery";
 export * from "./GalleryScene";
 export * from "./GalleryItem";
+export * from "./SolidColorItem";
+export * from "./ImageItem";
+export * from "./VideoItem";
